pyproject.toml
[project]
name = "igcheck"
version = "0.1.0"
description = "CLI tool to find Instagram accounts you follow that don't follow you back"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "instagrapi>=2.0.0",
  "click>=8.0.0",
  "rich>=13.0.0",
  "python-dotenv>=1.0.0",
  "questionary>=2.0.0",
]

[project.scripts]
igcheck = "igcheck.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
